+++ common/src/main/kotlin/com/cobblemon/mod/common/client/gui/battle/subscreen/BattleTargetSelectionPortrait.kt
@@ -118,7 +118,7 @@ class BattleTargetSelectionPortrait(
         backButton.render(context, mouseX, mouseY, delta)
     }
 
-    override fun mouseClicked(mouseX: Double, mouseY: Double, button: Int): Boolean {
+    override fun mousePrimaryClicked(mouseX: Double, mouseY: Double): Boolean {
         val target = targetTiles.find { it.isHovered(mouseX, mouseY) }
         if (target != null) {
             target.onClick()